refactor : apply separated components
